chore: 범위에 맞게 조정

+++ src/260111/ComeBackHome.kt
@@ -4,7 +4,7 @@ import kotlin.math.min
 import kotlin.math.sqrt
 
 fun main() {
-    val (X, Y, D, T) = readln().split(" ").map { it.toInt() }
+    val (X, Y, D, T) = readln().split(" ").map { it.toLong() }
 
     val distance = sqrt(X * X + Y * Y + 0.0)
